Extract formula desc auditing code to a separate module
The extracted module will be used for Cask descs as well.
