+++ Library/Homebrew/rubocops/extend/formula.rb
@@ -10,11 +10,13 @@ ensure
 end
 
 require "extend/string"
+require "rubocops/shared/helper_functions"
 
 module RuboCop
   module Cop
     class FormulaCop < Cop
       include RangeHelp
+      include HelperFunctions
 
       attr_accessor :file_path
 
@@ -30,28 +32,6 @@ module RuboCop
         class_node, parent_class_node, @body = *node
         @formula_name = Pathname.new(@file_path).basename(".rb").to_s
         audit_formula(node, class_node, parent_class_node, @body)
-      end
-
-      # Checks for regex match of pattern in the node and
-      # sets the appropriate instance variables to report the match
-      def regex_match_group(node, pattern)
-        string_repr = string_content(node).encode("UTF-8", invalid: :replace)
-        match_object = string_repr.match(pattern)
-        return unless match_object
-
-        node_begin_pos = start_column(node)
-        line_begin_pos = line_start_column(node)
-        @column = if node_begin_pos == line_begin_pos
-          node_begin_pos + match_object.begin(0) - line_begin_pos
-        else
-          node_begin_pos + match_object.begin(0) - line_begin_pos + 1
-        end
-        @length = match_object.to_s.length
-        @line_no = line_number(node)
-        @source_buf = source_buffer(node)
-        @offense_source_range = source_range(@source_buf, @line_no, @column, @length)
-        @offensive_node = node
-        match_object
       end
 
       # Yields to block when there is a match.
@@ -442,24 +422,9 @@ module RuboCop
         end
       end
 
-      # Returns the begin position of the node's line in source code
-      def line_start_column(node)
-        node.source_range.source_buffer.line_range(node.loc.line).begin_pos
-      end
-
-      # Returns the begin position of the node in source code
-      def start_column(node)
-        node.source_range.begin_pos
-      end
-
       # Returns the ending position of the node in source code
       def end_column(node)
         node.source_range.end_pos
-      end
-
-      # Returns the line number of the node
-      def line_number(node)
-        node.loc.line
       end
 
       # Returns the class node's name, nil if not a class node
@@ -484,35 +449,6 @@ module RuboCop
         block.loc.end.line - block.loc.begin.line
       end
 
-      # Source buffer is required as an argument to report style violations
-      def source_buffer(node)
-        node.source_range.source_buffer
-      end
-
-      # Returns the string representation if node is of type str(plain) or dstr(interpolated) or const
-      def string_content(node)
-        case node.type
-        when :str
-          node.str_content
-        when :dstr
-          content = ""
-          node.each_child_node(:str, :begin) do |child|
-            content += if child.begin_type?
-              child.source
-            else
-              child.str_content
-            end
-          end
-          content
-        when :const
-          node.const_name
-        when :sym
-          node.children.first.to_s
-        else
-          ""
-        end
-      end
-
       # Returns true if the formula is versioned
       def versioned_formula?
         @formula_name.include?("@")
@@ -530,10 +466,6 @@ module RuboCop
         return unless match_obj = @file_path.match(%r{/(homebrew-\w+)/})
 
         match_obj[1]
-      end
-
-      def problem(msg)
-        add_offense(@offensive_node, location: @offense_source_range, message: msg)
       end
 
       private
